+++ modules/hanna-css/src/lib/style-server-info.ts
@@ -79,6 +79,7 @@ export type CssModuleToken =
   | 'PullQuote'
   | 'RadioButtonsGroup'
   | 'RadioGroup'
+  | 'ReadSpeakerPlayer'
   | 'RelatedLinks'
   | 'RowBlock'
   | 'RowBlockColumn'